The ESEA analysis

the Extended-data Set Enrichment Analysis which includes gene and metabolite

Usage

ESEA(
  Ranks_all,
  minSize = 5,
  nPermSimple = 20000,
  gseaParam = 0.5,
  out = "Extended"
)

Arguments

Ranks_all

the Named vector of compound-level stats or gene-level stats. Names should be KEGG ID and (or) gene symbol.

minSize

Minimal size of a compound and(or) gene set to test. All pathways below the threshold are excluded.

nPermSimple

nPermSimple is the number of permutations used to estimate the significance of the enrichment score. The default value is 20000.

gseaParam

GSEA parameter value, all compound-level and(or) gene-level stats are raised to the power of 'gseaParam' before calculation of GSEA enrichment scores.

out

The pathway type for gene or metabolite,or extended pathway included genes and metabolites,default is "Extended",alternative is "metabolite" and "gene"

Value

the ESEA result

Examples

result <- ESEA(sim.cpd.data, out="metabolite")
result
#>                                                 pathway        pval      padj
#>                                                  <char>       <num>     <num>
#>  1:         Alanine, aspartate and glutamate metabolism 0.974672139 0.9863837
#>  2:         Amino sugar and nucleotide sugar metabolism 0.081894425 0.6858658
#>  3:                         Arachidonic acid metabolism 0.227226721 0.8975323
#>  4:                     Arginine and proline metabolism 0.329038149 0.8975323
#>  5:                               Arginine biosynthesis 0.986383660 0.9863837
#>  6:                   Ascorbate and aldarate metabolism 0.830520465 0.9449554
#>  7:             Biosynthesis of unsaturated fatty acids 0.007316829 0.4902275
#>  8:                                   Biotin metabolism 0.483135916 0.9449554
#>  9:                                Butanoate metabolism 0.018158236 0.6083009
#> 10:                                 Caffeine metabolism 0.836077093 0.9449554
#> 11:                           Citrate cycle (TCA cycle) 0.944333400 0.9863837
#> 12:                  Cysteine and methionine metabolism 0.569176352 0.9449554
#> 13:                             D-Amino acid metabolism 0.388484113 0.8975323
#> 14:                   Drug metabolism - cytochrome P450 0.255327664 0.8975323
#> 15:                     Drug metabolism - other enzymes 0.221367180 0.8975323
#> 16:                              Ether lipid metabolism 0.863735827 0.9449554
#> 17:                             Fatty acid biosynthesis 0.517569326 0.9449554
#> 18:                              Fatty acid degradation 0.677311256 0.9449554
#> 19:                               Fatty acid elongation 0.686278440 0.9449554
#> 20:                                 Folate biosynthesis 0.297775129 0.8975323
#> 21:                     Fructose and mannose metabolism 0.821533773 0.9449554
#> 22:                                Galactose metabolism 0.319447187 0.8975323
#> 23:                              Glutathione metabolism 0.064896166 0.6211490
#> 24:                             Glycerolipid metabolism 0.698366570 0.9449554
#> 25:                      Glycerophospholipid metabolism 0.914041268 0.9720756
#> 26:            Glycine, serine and threonine metabolism 0.386176799 0.8975323
#> 27:                        Glycolysis / Gluconeogenesis 0.790269296 0.9449554
#> 28:             Glyoxylate and dicarboxylate metabolism 0.567131691 0.9449554
#> 29:                                Histidine metabolism 0.028310321 0.6211490
#> 30:                       Inositol phosphate metabolism 0.376623377 0.8975323
#> 31:                            Linoleic acid metabolism 0.665705479 0.9449554
#> 32:                                  Lysine degradation 0.968232044 0.9863837
#> 33:        Metabolism of xenobiotics by cytochrome P450 0.778090452 0.9449554
#> 34:     Neomycin, kanamycin and gentamicin biosynthesis 0.570685343 0.9449554
#> 35:              Nicotinate and nicotinamide metabolism 0.352663188 0.8975323
#> 36:                                 Nitrogen metabolism 0.493188824 0.9449554
#> 37:                           Oxidative phosphorylation 0.432172869 0.9449554
#> 38:                   Pantothenate and CoA biosynthesis 0.759947880 0.9449554
#> 39:            Pentose and glucuronate interconversions 0.683303085 0.9449554
#> 40:                           Pentose phosphate pathway 0.200059073 0.8975323
#> 41:                            Phenylalanine metabolism 0.371870324 0.8975323
#> 42: Phenylalanine, tyrosine and tryptophan biosynthesis 0.674772341 0.9449554
#> 43:              Phosphonate and phosphinate metabolism 0.869538957 0.9449554
#> 44:                                Porphyrin metabolism 0.843423588 0.9449554
#> 45:                      Primary bile acid biosynthesis 0.221254878 0.8975323
#> 46:                               Propanoate metabolism 0.266134663 0.8975323
#> 47:                                   Purine metabolism 0.159090909 0.8975323
#> 48:                               Pyrimidine metabolism 0.556591608 0.9449554
#> 49:                                 Pyruvate metabolism 0.062068275 0.6211490
#> 50:                                  Retinol metabolism 0.776494024 0.9449554
#> 51:                               Riboflavin metabolism 0.825179712 0.9449554
#> 52:                           Selenocompound metabolism 0.264241455 0.8975323
#> 53:                             Sphingolipid metabolism 0.683210662 0.9449554
#> 54:                       Starch and sucrose metabolism 0.811666834 0.9449554
#> 55:                                Steroid biosynthesis 0.266212288 0.8975323
#> 56:                        Steroid hormone biosynthesis 0.038150172 0.6211490
#> 57:                                   Sulfur metabolism 0.128008753 0.8975323
#> 58:                  Taurine and hypotaurine metabolism 0.462755306 0.9449554
#> 59:                     Terpenoid backbone biosynthesis 0.265632816 0.8975323
#> 60:                                 Thiamine metabolism 0.582010050 0.9449554
#> 61:                               Tryptophan metabolism 0.792331953 0.9449554
#> 62:                                 Tyrosine metabolism 0.761900040 0.9449554
#> 63: Ubiquinone and other terpenoid-quinone biosynthesis 0.059629815 0.6211490
#> 64:         Valine, leucine and isoleucine biosynthesis 0.328274198 0.8975323
#> 65:          Valine, leucine and isoleucine degradation 0.874436316 0.9449554
#> 66:                               Vitamin B6 metabolism 0.458578995 0.9449554
#> 67:                             beta-Alanine metabolism 0.269825436 0.8975323
#>                                                 pathway        pval      padj
#>        log2err         ES        NES  size
#>          <num>      <num>      <num> <int>
#>  1: 0.01047501  0.1603305  0.5279825    11
#>  2: 0.04904201  0.2788773  1.4066839    38
#>  3: 0.02869240 -0.2422494 -1.1968733    35
#>  4: 0.02319131 -0.2481361 -1.1004331    25
#>  5: 0.01035415  0.1547907  0.4932924    10
#>  6: 0.01221553 -0.1603550 -0.7210450    26
#>  7: 0.16912722 -0.4807547 -1.8672289    17
#>  8: 0.01812254  0.3714095  0.9690910     6
#>  9: 0.10659973  0.4821621  1.7380507    14
#> 10: 0.01207656  0.2665006  0.6953601     6
#> 11: 0.01078053 -0.2434651 -0.5930879     5
#> 12: 0.01629504 -0.2038355 -0.9165574    26
#> 13: 0.02093249 -0.2250702 -1.0484679    29
#> 14: 0.02667728  0.3387346  1.1860269    13
#> 15: 0.02891049 -0.5094060 -1.2409276     5
#> 16: 0.01164588  0.2295066  0.6726305     8
#> 17: 0.01727886  0.2870615  0.9453189    11
#> 18: 0.01410081  0.1987853  0.8332429    21
#> 19: 0.01413998 -0.2172983 -0.8257829    16
#> 20: 0.02429614  0.3347711  1.1389362    12
#> 21: 0.01238105 -0.1771190 -0.7263324    20
#> 22: 0.02322123  0.2563854  1.1054502    23
#> 23: 0.05568218 -0.4748371 -1.5202730    10
#> 24: 0.01395051 -0.2258991 -0.8171760    14
#> 25: 0.01118692 -0.1590514 -0.6297799    18
#> 26: 0.02104792 -0.2401409 -1.0497276    24
#> 27: 0.01263390  0.2244597  0.7391656    11
#> 28: 0.01609595  0.2051749  0.9199673    26
#> 29: 0.08492238  0.4173039  1.6545911    18
#> 30: 0.02137109 -0.2535908 -1.0582263    21
#> 31: 0.01438198  0.2972107  0.8251728     7
#> 32: 0.01039358  0.1350773  0.5472927    19
#> 33: 0.01283716 -0.2438117 -0.7479766     9
#> 34: 0.01615124  0.2592160  0.9076048    13
#> 35: 0.02184213  0.2618361  1.0814056    20
#> 36: 0.01776534  0.3459865  0.9605934     7
#> 37: 0.01943678  0.4146070  1.0061896     5
#> 38: 0.01305629 -0.2032752 -0.7724919    16
#> 39: 0.01425059 -0.1839892 -0.8379946    27
#> 40: 0.03033945  0.2999153  1.2386760    20
#> 41: 0.02131469  0.2809576  1.0665963    16
#> 42: 0.01430483  0.2252883  0.8320884    15
#> 43: 0.01170056 -0.1957246 -0.6705056    12
#> 44: 0.01209148 -0.1391971 -0.7328997    43
#> 45: 0.02893989  0.3289827  1.2150771    15
#> 46: 0.02600392  0.2994309  1.1614823    17
#> 47: 0.03440220  0.2499584  1.2732934    39
#> 48: 0.01633743  0.1987120  0.9328600    30
#> 49: 0.05705865  0.4627760  1.5239618    11
#> 50: 0.01276442 -0.2848485 -0.7476039     6
#> 51: 0.01215642 -0.2214620 -0.7090487    10
#> 52: 0.02613977 -0.4851419 -1.1818195     5
#> 53: 0.01410550  0.2762005  0.8094795     8
#> 54: 0.01236821 -0.1916463 -0.7282995    16
#> 55: 0.02607163 -0.3061291 -1.1633603    16
#> 56: 0.07364873 -0.3392886 -1.5805432    29
#> 57: 0.03890743  0.4698493  1.3770191     8
#> 58: 0.01860916  0.3096990  0.9869595    10
#> 59: 0.02607729  0.3347635  1.1721227    13
#> 60: 0.01598060 -0.2921486 -0.8962665     9
#> 61: 0.01244424  0.1780306  0.7462461    21
#> 62: 0.01291734  0.1732044  0.7857134    27
#> 63: 0.05823311  0.4339129  1.5192788    13
#> 64: 0.02309363 -0.3787439 -1.1110849     8
#> 65: 0.01159624 -0.1850583 -0.6694369    14
#> 66: 0.01869598 -0.2807600 -0.9898141    13
#> 67: 0.02579785  0.3064691  1.1634453    16
#>        log2err         ES        NES  size
#>                                                                                                                                                                                                                leadingEdge
#>                                                                                                                                                                                                                     <char>
#>  1:                                                                                                                                                                                                 C00042, C00014, C00049
#>  2:                                                                                                                                         C00159, C00734, C00645, C00357, C16698, C00461, C00031, C01170, C02280, C03733
#>  3:                                                                                                                                                                 C14795, C05966, C05955, C00219, C04758, C04835, C14775
#>  4:                                                                                                                                         C01682, C04281, C00077, C00750, C05147, C00581, C05945, C00431, C01250, C00406
#>  5:                                                                                                                                                                                                 C00014, C00437, C00049
#>  6:                                                                                                                                                                 C01114, C03921, C00679, C01115, C00333, C00558, C01146
#>  7:                                                                                                                                                 C08323, C02050, C16166, C00219, C00412, C16529, C00154, C16513, C16522
#>  8:                                                                                                                                                                         C00047, C00120, C01063, C01092, C05921, C01037
#>  9:                                                                                                                                                                         C00246, C06146, C00132, C03561, C06142, C00042
#> 10:                                                                                                                                                                                                                 C16352
#> 11:                                                                                                                                                                                                 C00417, C00036, C00026
#> 12:                                                                                                                 C05527, C03232, C08276, C02218, C01005, C03082, C15651, C00021, C05324, C04188, C05528, C01180, C00197
#> 13:                                                                                                                                                                         C01726, C04282, C00819, C03239, C00077, C02855
#> 14:                                                                                                                                                                 C07185, C01516, C16561, C07647, C07643, C07572, C01471
#> 15:                                                                                                                                                                                         C00003, C00006, C16633, C11736
#> 16:                                                                                                                                                                                 C04598, C01264, C04517, C03454, C05212
#> 17:                                                                                                                                                                         C00383, C00249, C01530, C05761, C08362, C01571
#> 18:                                                                                                                                                                                         C05102, C03561, C05268, C00249
#> 19:                                                                                                                                                                                 C05272, C05260, C00154, C00040, C05258
#> 20:                                                                                                                                                                         C00251, C00272, C00921, C06313, C03541, C09332
#> 21:                                                                                                                                         C00111, C01131, C00275, C00247, C11516, C01768, C02977, C00636, C01222, C03267
#> 22:                                                                                                                                                                                 C00159, C06376, C00031, C03733, C05402
#> 23:                                                                                                                                                                                 C05730, C00077, C00750, C00006, C03646
#> 24:                                                                                                                                                                 C00111, C06037, C00969, C00040, C00197, C02133, C00631
#> 25:                                                                                                                                                                                 C04438, C00111, C03974, C00040, C04638
#> 26:                                                                                                                                                         C00576, C00101, C03232, C00143, C05519, C01005, C03082, C00581
#> 27:                                                                                                                                                                                                         C05378, C00031
#> 28:                                                                                                                                                                         C06027, C00258, C03561, C00042, C00014, C00100
#> 29:                                                                                                                                         C05575, C00785, C01152, C03680, C05568, C05827, C02835, C00049, C01045, C04896
#> 30:                                                                                                                                                         C01272, C03659, C00111, C00691, C03546, C00558, C04287, C06153
#> 31:                                                                                                                                                                                                 C07289, C14762, C14767
#> 32:                                                                                                                                                                                                 C05825, C00042, C00047
#> 33:                                                                                                                                                                                                         C14039, C00745
#> 34:                                                                                                                                                                         C00031, C07656, C00908, C17580, C00029, C00832
#> 35:                                                                                                                                         C04226, C01185, C03458, C00042, C03652, C00049, C06178, C01020, C00163, C05842
#> 36:                                                                                                                                                                                                         C01563, C00014
#> 37:                                                                                                                                                                                                 C00042, C00001, C00061
#> 38:                                                                                                                                                                                 C02642, C00750, C00183, C00010, C05944
#> 39: C01904, C00111, C04053, C00333, C00558, C00181, C00309, C00266, C00817, C00026, C14899, C02753, C00379, C03291, C00474, C01101, C00476, C00800, C00191, C02266, C00231, C01068, C00514, C00029, C00905, C01508, C04575
#> 40:                                                                                                                                                         C05382, C00258, C05378, C00031, C00672, C05345, C00279, C00257
#> 41:                                                                                                                                                                                 C00082, C02505, C00042, C00166, C11588
#> 42:                                                                                                                                                                                         C00251, C00082, C00296, C00166
#> 43:                                                                                                                                                                                                         C06451, C01151
#> 44:                         C06399, C01079, C00500, C05794, C03373, C01708, C11543, C05793, C03263, C05768, C11630, C05774, C05797, C03114, C03741, C05913, C00037, C06408, C11831, C05912, C05775, C05773, C06505, C05770
#> 45:                                                                                                                                                                                         C06341, C05454, C05122, C15519
#> 46:                                                                                                                                                 C06002, C00546, C00042, C00100, C00894, C00099, C00163, C00804, C01213
#> 47:                                                                                                                 C05515, C00575, C00014, C05513, C04640, C00802, C00655, C01345, C05516, C00104, C04734, C06197, C00053
#> 48:                                                                                                                                                                         C00383, C01563, C00429, C01205, C00526, C00075
#> 49:                                                                                                                                                                                 C00546, C00227, C00042, C03451, C03981
#> 50:                                                                                                                                                                                                                 C11061
#> 51:                                                                                                                                                                                                                 C00235
#> 52:                                                                                                                                                                                 C05697, C05688, C05336, C05699, C05172
#> 53:                                                                                                                                                                                                                 C03640
#> 54:                                                                                                                                                                                 C00498, C02591, C00085, C01898, C03323
#> 55:                                                                                                                                                         C01164, C05442, C11455, C01054, C11508, C05437, C11523, C00751
#> 56:                                                                                                                         C02537, C03852, C11134, C05498, C00735, C02538, C03917, C02373, C03772, C05485, C02140, C05480
#> 57:                                                                                                                                                                                         C00042, C01861, C00283, C00053
#> 58:                                                                                                                                         C00227, C05122, C14179, C03149, C00506, C00033, C05123, C01678, C00026, C00593
#> 59:                                                                                                                                                                                 C00353, C11434, C11437, C11435, C04574
#> 60:                                                                                                                                                                                                 C00003, C04752, C03373
#> 61:                                                                                                                                                         C00978, C05658, C08313, C03230, C02298, C00331, C01598, C00780
#> 62:                                                                                                                                                                                 C00082, C03964, C05338, C00042, C00547
#> 63:                                                                                                                                                                                 C00251, C00353, C00082, C03964, C16695
#> 64:                                                                                                                                                                 C06032, C00183, C00671, C04272, C00233, C02226, C14463
#> 65:                                                                                                                                                                         C02170, C05998, C00183, C06001, C00671, C00683
#> 66:                                                                                                                                                                         C00534, C00647, C00232, C07335, C00266, C01215
#> 67:                                                                                                                                                                 C00383, C00429, C00864, C00100, C00894, C00099, C00049
#>                                                                                                                                                                                                                leadingEdge