Feature selection using Boruta

Feature selection using Boruta

Usage

ML_Boruta(object)

Arguments

object

A dataframe-like data object containing log-metabolite intensity values, with columns corresponding to metabolites and must containing the group column, and the rows corresponding to the samples

Value

test

Examples

library(dplyr)

meta_dat1 <- t(meta_dat) %>%
  as.data.frame() %>%
  dplyr::mutate(group=group)

result_ML_Boruta <- ML_Boruta(meta_dat1)
#>  1. run of importance source...
#>  2. run of importance source...
#>  3. run of importance source...
#>  4. run of importance source...
#>  5. run of importance source...
#>  6. run of importance source...
#>  7. run of importance source...
#>  8. run of importance source...
#>  9. run of importance source...
#>  10. run of importance source...
#>  11. run of importance source...
#>  12. run of importance source...
#>  13. run of importance source...
#>  14. run of importance source...
#>  15. run of importance source...
#> After 15 iterations, +0.51 secs: 
#>  confirmed 17 attributes: C00022, C00025, C00042, C00073, C00148 and 12 more;
#>  rejected 121 attributes: C00002, C00003, C00006, C00008, C00015 and 116 more;
#>  still have 81 attributes left.
#>  16. run of importance source...
#>  17. run of importance source...
#>  18. run of importance source...
#>  19. run of importance source...
#> After 19 iterations, +0.6 secs: 
#>  confirmed 3 attributes: C00079, C00170, C02985;
#>  rejected 19 attributes: C00004, C00020, C00044, C00064, C00078 and 14 more;
#>  still have 59 attributes left.
#>  20. run of importance source...
#>  21. run of importance source...
#>  22. run of importance source...
#>  23. run of importance source...
#> After 23 iterations, +0.71 secs: 
#>  confirmed 3 attributes: C00307, C03413, C03793;
#>  rejected 4 attributes: C00055, C00183, C01210, C01601;
#>  still have 52 attributes left.
#>  24. run of importance source...
#>  25. run of importance source...
#>  26. run of importance source...
#> After 26 iterations, +0.77 secs: 
#>  confirmed 1 attribute: C00570;
#>  rejected 5 attributes: C00035, C00105, C00188, C00295, C02862;
#>  still have 46 attributes left.
#>  27. run of importance source...
#>  28. run of importance source...
#>  29. run of importance source...
#>  30. run of importance source...
#> After 30 iterations, +0.85 secs: 
#>  confirmed 2 attributes: C00149, C00447;
#>  still have 44 attributes left.
#>  31. run of importance source...
#>  32. run of importance source...
#>  33. run of importance source...
#> After 33 iterations, +0.91 secs: 
#>  confirmed 4 attributes: C00153, C00167, C00490, C06192;
#>  rejected 1 attribute: C00152;
#>  still have 39 attributes left.
#>  34. run of importance source...
#>  35. run of importance source...
#>  36. run of importance source...
#> After 36 iterations, +0.97 secs: 
#>  rejected 1 attribute: C16513;
#>  still have 38 attributes left.
#>  37. run of importance source...
#>  38. run of importance source...
#>  39. run of importance source...
#> After 39 iterations, +1 secs: 
#>  rejected 2 attributes: C00408, C00624;
#>  still have 36 attributes left.
#>  40. run of importance source...
#>  41. run of importance source...
#>  42. run of importance source...
#> After 42 iterations, +1.1 secs: 
#>  confirmed 1 attribute: C03406;
#>  still have 35 attributes left.
#>  43. run of importance source...
#>  44. run of importance source...
#>  45. run of importance source...
#>  46. run of importance source...
#>  47. run of importance source...
#>  48. run of importance source...
#> After 48 iterations, +1.2 secs: 
#>  rejected 1 attribute: C00438;
#>  still have 34 attributes left.
#>  49. run of importance source...
#>  50. run of importance source...
#> After 50 iterations, +1.3 secs: 
#>  confirmed 1 attribute: C00013;
#>  rejected 1 attribute: C00385;
#>  still have 32 attributes left.
#>  51. run of importance source...
#>  52. run of importance source...
#>  53. run of importance source...
#>  54. run of importance source...
#>  55. run of importance source...
#>  56. run of importance source...
#>  57. run of importance source...
#>  58. run of importance source...
#>  59. run of importance source...
#>  60. run of importance source...
#>  61. run of importance source...
#>  62. run of importance source...
#>  63. run of importance source...
#>  64. run of importance source...
#>  65. run of importance source...
#>  66. run of importance source...
#>  67. run of importance source...
#>  68. run of importance source...
#>  69. run of importance source...
#>  70. run of importance source...
#>  71. run of importance source...
#>  72. run of importance source...
#>  73. run of importance source...
#>  74. run of importance source...
#>  75. run of importance source...
#>  76. run of importance source...
#>  77. run of importance source...
#> After 77 iterations, +1.8 secs: 
#>  rejected 1 attribute: C00588;
#>  still have 31 attributes left.
#>  78. run of importance source...
#>  79. run of importance source...
#>  80. run of importance source...
#> After 80 iterations, +1.8 secs: 
#>  rejected 1 attribute: C00065;
#>  still have 30 attributes left.
#>  81. run of importance source...
#>  82. run of importance source...
#>  83. run of importance source...
#>  84. run of importance source...
#>  85. run of importance source...
#> After 85 iterations, +1.9 secs: 
#>  confirmed 1 attribute: C00010;
#>  still have 29 attributes left.
#>  86. run of importance source...
#>  87. run of importance source...
#>  88. run of importance source...
#>  89. run of importance source...
#>  90. run of importance source...
#>  91. run of importance source...
#>  92. run of importance source...
#>  93. run of importance source...
#>  94. run of importance source...
#>  95. run of importance source...
#>  96. run of importance source...
#>  97. run of importance source...
#>  98. run of importance source...
#>  99. run of importance source...
#>  100. run of importance source...
#>  101. run of importance source...
#>  102. run of importance source...
#>  103. run of importance source...
#>  104. run of importance source...
#>  105. run of importance source...
#> After 105 iterations, +2.3 secs: 
#>  confirmed 1 attribute: C00097;
#>  still have 28 attributes left.
#>  106. run of importance source...
#>  107. run of importance source...
#>  108. run of importance source...
#>  109. run of importance source...
#>  110. run of importance source...
#>  111. run of importance source...
#>  112. run of importance source...
#> After 112 iterations, +2.5 secs: 
#>  rejected 1 attribute: C01419;
#>  still have 27 attributes left.
#>  113. run of importance source...
#>  114. run of importance source...
#>  115. run of importance source...
#>  116. run of importance source...
#>  117. run of importance source...
#>  118. run of importance source...
#>  119. run of importance source...
#>  120. run of importance source...
#>  121. run of importance source...
#>  122. run of importance source...
#>  123. run of importance source...
#>  124. run of importance source...
#>  125. run of importance source...
#> After 125 iterations, +2.7 secs: 
#>  confirmed 1 attribute: C00352;
#>  still have 26 attributes left.
#>  126. run of importance source...
#>  127. run of importance source...
#>  128. run of importance source...
#>  129. run of importance source...
#>  130. run of importance source...
#>  131. run of importance source...
#>  132. run of importance source...
#>  133. run of importance source...
#>  134. run of importance source...
#> After 134 iterations, +2.9 secs: 
#>  confirmed 1 attribute: C05673;
#>  still have 25 attributes left.
#>  135. run of importance source...
#>  136. run of importance source...
#>  137. run of importance source...
#>  138. run of importance source...
#>  139. run of importance source...
#>  140. run of importance source...
#>  141. run of importance source...
#>  142. run of importance source...
#>  143. run of importance source...
#>  144. run of importance source...
#>  145. run of importance source...
#>  146. run of importance source...
#>  147. run of importance source...
#>  148. run of importance source...
#>  149. run of importance source...
#>  150. run of importance source...
#>  151. run of importance source...
#>  152. run of importance source...
#>  153. run of importance source...
#>  154. run of importance source...
#>  155. run of importance source...
#>  156. run of importance source...
#>  157. run of importance source...
#>  158. run of importance source...
#>  159. run of importance source...
#>  160. run of importance source...
#>  161. run of importance source...
#>  162. run of importance source...
#>  163. run of importance source...
#>  164. run of importance source...
#>  165. run of importance source...
#>  166. run of importance source...
#>  167. run of importance source...
#>  168. run of importance source...
#> After 168 iterations, +3.6 secs: 
#>  confirmed 1 attribute: C00670;
#>  still have 24 attributes left.
#>  169. run of importance source...
#>  170. run of importance source...
#>  171. run of importance source...
#>  172. run of importance source...
#>  173. run of importance source...
#>  174. run of importance source...
#>  175. run of importance source...
#>  176. run of importance source...
#>  177. run of importance source...
#>  178. run of importance source...
#>  179. run of importance source...
#>  180. run of importance source...
#>  181. run of importance source...
#>  182. run of importance source...
#>  183. run of importance source...
#>  184. run of importance source...
#> After 184 iterations, +3.9 secs: 
#>  rejected 1 attribute: C00186;
#>  still have 23 attributes left.
#>  185. run of importance source...
#>  186. run of importance source...
#>  187. run of importance source...
#>  188. run of importance source...
#>  189. run of importance source...
#>  190. run of importance source...
#>  191. run of importance source...
#>  192. run of importance source...
#>  193. run of importance source...
#>  194. run of importance source...
#>  195. run of importance source...
#>  196. run of importance source...
#>  197. run of importance source...
#>  198. run of importance source...
#>  199. run of importance source...
#>  200. run of importance source...
#>  201. run of importance source...
#>  202. run of importance source...
#>  203. run of importance source...
#>  204. run of importance source...
#>  205. run of importance source...
#>  206. run of importance source...
#>  207. run of importance source...
#>  208. run of importance source...
#>  209. run of importance source...
#>  210. run of importance source...
#>  211. run of importance source...
#>  212. run of importance source...
#>  213. run of importance source...
#>  214. run of importance source...
#> After 214 iterations, +4.5 secs: 
#>  rejected 1 attribute: C06241;
#>  still have 22 attributes left.
#>  215. run of importance source...
#>  216. run of importance source...
#>  217. run of importance source...
#>  218. run of importance source...
#>  219. run of importance source...
#>  220. run of importance source...
#>  221. run of importance source...
#>  222. run of importance source...
#>  223. run of importance source...
#>  224. run of importance source...
#>  225. run of importance source...
#>  226. run of importance source...
#>  227. run of importance source...
#>  228. run of importance source...
#>  229. run of importance source...
#>  230. run of importance source...
#>  231. run of importance source...
#>  232. run of importance source...
#>  233. run of importance source...
#>  234. run of importance source...
#>  235. run of importance source...
#> After 235 iterations, +4.9 secs: 
#>  rejected 1 attribute: C00041;
#>  still have 21 attributes left.
#>  236. run of importance source...
#>  237. run of importance source...
#>  238. run of importance source...
#>  239. run of importance source...
#>  240. run of importance source...
#>  241. run of importance source...
#>  242. run of importance source...
#>  243. run of importance source...
#>  244. run of importance source...
#>  245. run of importance source...
#>  246. run of importance source...
#>  247. run of importance source...
#>  248. run of importance source...
#>  249. run of importance source...
#>  250. run of importance source...
#>  251. run of importance source...
#>  252. run of importance source...
#>  253. run of importance source...
#>  254. run of importance source...
#>  255. run of importance source...
#>  256. run of importance source...
#>  257. run of importance source...
#>  258. run of importance source...
#>  259. run of importance source...
#>  260. run of importance source...
#>  261. run of importance source...
#>  262. run of importance source...
#>  263. run of importance source...
#>  264. run of importance source...
#>  265. run of importance source...
#>  266. run of importance source...
#>  267. run of importance source...
#>  268. run of importance source...
#>  269. run of importance source...
#>  270. run of importance source...
#>  271. run of importance source...
#>  272. run of importance source...
#>  273. run of importance source...
#>  274. run of importance source...
#>  275. run of importance source...
#>  276. run of importance source...
#>  277. run of importance source...
#>  278. run of importance source...
#>  279. run of importance source...
#>  280. run of importance source...
#>  281. run of importance source...
#>  282. run of importance source...
#>  283. run of importance source...
#>  284. run of importance source...
#>  285. run of importance source...
#>  286. run of importance source...
#>  287. run of importance source...
#>  288. run of importance source...
#>  289. run of importance source...
#>  290. run of importance source...
#>  291. run of importance source...
#>  292. run of importance source...
#>  293. run of importance source...
#>  294. run of importance source...
#>  295. run of importance source...
#>  296. run of importance source...
#>  297. run of importance source...
#>  298. run of importance source...
#>  299. run of importance source...
#>  300. run of importance source...
#>  301. run of importance source...
#>  302. run of importance source...
#>  303. run of importance source...
#> After 303 iterations, +6.2 secs: 
#>  confirmed 1 attribute: C13482;
#>  still have 20 attributes left.
#>  304. run of importance source...
#>  305. run of importance source...
#>  306. run of importance source...
#>  307. run of importance source...
#>  308. run of importance source...
#>  309. run of importance source...
#>  310. run of importance source...
#>  311. run of importance source...
#>  312. run of importance source...
#>  313. run of importance source...
#>  314. run of importance source...
#>  315. run of importance source...
#>  316. run of importance source...
#>  317. run of importance source...
#>  318. run of importance source...
#>  319. run of importance source...
#>  320. run of importance source...
#>  321. run of importance source...
#>  322. run of importance source...
#>  323. run of importance source...
#>  324. run of importance source...
#>  325. run of importance source...
#>  326. run of importance source...
#>  327. run of importance source...
#>  328. run of importance source...
#>  329. run of importance source...
#>  330. run of importance source...
#>  331. run of importance source...
#>  332. run of importance source...
#>  333. run of importance source...
#>  334. run of importance source...
#>  335. run of importance source...
#>  336. run of importance source...
#>  337. run of importance source...
#>  338. run of importance source...
#>  339. run of importance source...
#>  340. run of importance source...
#>  341. run of importance source...
#>  342. run of importance source...
#>  343. run of importance source...
#>  344. run of importance source...
#>  345. run of importance source...
#>  346. run of importance source...
#>  347. run of importance source...
#>  348. run of importance source...
#>  349. run of importance source...
#>  350. run of importance source...
#>  351. run of importance source...
#>  352. run of importance source...
#>  353. run of importance source...
#>  354. run of importance source...
#>  355. run of importance source...
#>  356. run of importance source...
#>  357. run of importance source...
#>  358. run of importance source...
#>  359. run of importance source...
#>  360. run of importance source...
#>  361. run of importance source...
#>  362. run of importance source...
#>  363. run of importance source...
#>  364. run of importance source...
#> After 364 iterations, +7.4 secs: 
#>  confirmed 1 attribute: C03264;
#>  still have 19 attributes left.
#>  365. run of importance source...
#>  366. run of importance source...
#>  367. run of importance source...
#>  368. run of importance source...
#>  369. run of importance source...
#>  370. run of importance source...
#>  371. run of importance source...
#>  372. run of importance source...
#>  373. run of importance source...
#>  374. run of importance source...
#>  375. run of importance source...
#>  376. run of importance source...
#>  377. run of importance source...
#>  378. run of importance source...
#>  379. run of importance source...
#>  380. run of importance source...
#>  381. run of importance source...
#>  382. run of importance source...
#>  383. run of importance source...
#>  384. run of importance source...
#>  385. run of importance source...
#>  386. run of importance source...
#>  387. run of importance source...
#>  388. run of importance source...
#>  389. run of importance source...
#>  390. run of importance source...
#>  391. run of importance source...
#>  392. run of importance source...
#>  393. run of importance source...
#>  394. run of importance source...
#>  395. run of importance source...
#>  396. run of importance source...
#>  397. run of importance source...
#>  398. run of importance source...
#>  399. run of importance source...
#> After 399 iterations, +8.1 secs: 
#>  rejected 1 attribute: C00051;
#>  still have 18 attributes left.
#>  400. run of importance source...
#>  401. run of importance source...
#>  402. run of importance source...
#>  403. run of importance source...
#>  404. run of importance source...
#>  405. run of importance source...
#>  406. run of importance source...
#>  407. run of importance source...
#>  408. run of importance source...
#>  409. run of importance source...
#>  410. run of importance source...
#>  411. run of importance source...
#>  412. run of importance source...
#>  413. run of importance source...
#>  414. run of importance source...
#>  415. run of importance source...
#>  416. run of importance source...
#>  417. run of importance source...
#>  418. run of importance source...
#>  419. run of importance source...
#>  420. run of importance source...
#>  421. run of importance source...
#>  422. run of importance source...
#>  423. run of importance source...
#>  424. run of importance source...
#>  425. run of importance source...
#>  426. run of importance source...
#>  427. run of importance source...
#>  428. run of importance source...
#>  429. run of importance source...
#>  430. run of importance source...
#>  431. run of importance source...
#>  432. run of importance source...
#>  433. run of importance source...
#>  434. run of importance source...
#>  435. run of importance source...
#> After 435 iterations, +8.8 secs: 
#>  confirmed 1 attribute: C00519;
#>  still have 17 attributes left.
#>  436. run of importance source...
#>  437. run of importance source...
#>  438. run of importance source...
#>  439. run of importance source...
#>  440. run of importance source...
#>  441. run of importance source...
#>  442. run of importance source...
#>  443. run of importance source...
#>  444. run of importance source...
#>  445. run of importance source...
#>  446. run of importance source...
#>  447. run of importance source...
#>  448. run of importance source...
#>  449. run of importance source...
#>  450. run of importance source...
#>  451. run of importance source...
#>  452. run of importance source...
#>  453. run of importance source...
#>  454. run of importance source...
#>  455. run of importance source...
#>  456. run of importance source...
#>  457. run of importance source...
#>  458. run of importance source...
#>  459. run of importance source...
#>  460. run of importance source...
#>  461. run of importance source...
#>  462. run of importance source...
#>  463. run of importance source...
#>  464. run of importance source...
#>  465. run of importance source...
#>  466. run of importance source...
#>  467. run of importance source...
#>  468. run of importance source...
#>  469. run of importance source...
#>  470. run of importance source...
#>  471. run of importance source...
#>  472. run of importance source...
#> After 472 iterations, +9.5 secs: 
#>  confirmed 2 attributes: C00106, C05851;
#>  still have 15 attributes left.
#>  473. run of importance source...
#>  474. run of importance source...
#>  475. run of importance source...
#>  476. run of importance source...
#>  477. run of importance source...
#>  478. run of importance source...
#>  479. run of importance source...
#>  480. run of importance source...
#>  481. run of importance source...
#>  482. run of importance source...
#>  483. run of importance source...
#>  484. run of importance source...
#>  485. run of importance source...
#>  486. run of importance source...
#>  487. run of importance source...
#>  488. run of importance source...
#>  489. run of importance source...
#>  490. run of importance source...
#>  491. run of importance source...
#>  492. run of importance source...
#>  493. run of importance source...
#>  494. run of importance source...
#>  495. run of importance source...
#>  496. run of importance source...
#>  497. run of importance source...
#>  498. run of importance source...
#>  499. run of importance source...
result_ML_Boruta
#>       name  meanImp medianImp      minImp   maxImp  normHits  decision
#> 1   C03264 1.830748  1.917029 -1.00100150 3.359104 0.5891784 Confirmed
#> 2   C02630 3.922494  3.964871  2.13766269 5.294886 1.0000000 Confirmed
#> 3   C00170 2.309354  2.366479 -1.00100150 3.490521 0.8577154 Confirmed
#> 4   C18170 3.553873  3.565182  1.94897610 4.802430 0.9959920 Confirmed
#> 5   C06192 2.264763  2.305483 -0.49248537 3.566407 0.8176353 Confirmed
#> 6   C00267 4.320160  4.331521  2.47752141 5.607688 1.0000000 Confirmed
#> 7   C05378 3.433145  3.444630  1.98073358 4.782041 0.9959920 Confirmed
#> 8   C00307 2.514800  2.555565  0.75593669 3.786108 0.9098196 Confirmed
#> 9   C00570 2.382516  2.435941  0.74948491 3.697595 0.8657315 Confirmed
#> 10  C05673 2.089739  2.161342 -1.00100150 3.717363 0.7274549 Confirmed
#> 11  C05674 4.244065  4.261572  2.48290780 5.632259 1.0000000 Confirmed
#> 12  C00010 2.208908  2.302146 -1.00100150 3.536697 0.7915832 Confirmed
#> 13  C05851 1.781306  1.888854 -1.00100150 3.139645 0.5891784 Confirmed
#> 14  C00352 1.968446  1.999450  0.07432983 3.221504 0.6793587 Confirmed
#> 15  C00013 2.500754  2.521148 -1.62224509 4.155475 0.8957916 Confirmed
#> 16  C00365 3.998296  4.025179  2.23682734 5.208333 1.0000000 Confirmed
#> 17  C00242 3.909177  3.884653  2.70881709 5.412645 1.0000000 Confirmed
#> 18  C00519 1.748288  1.894215 -1.00100150 3.133966 0.5871743 Confirmed
#> 19  C00490 2.551456  2.583895  1.00100150 3.794342 0.9018036 Confirmed
#> 20 C05938  3.840038  3.865926  2.15091351 4.950098 1.0000000 Confirmed
#> 21  C02291 4.348369  4.363354  2.81417060 6.012771 1.0000000 Confirmed
#> 22  C00097 2.112802  2.146414 -0.44305251 3.776334 0.7595190 Confirmed
#> 23  C02045 5.155510  5.169558  3.68291606 6.756627 1.0000000 Confirmed
#> 24  C02985 3.716313  3.728260  1.73095465 5.192007 0.9959920 Confirmed
#> 25  C00025 3.853282  3.812617  2.52640808 5.315595 1.0000000 Confirmed
#> 26  C00407 3.514677  3.560238  1.93039736 4.859897 0.9979960 Confirmed
#> 27  C00073 5.160100  5.188471  3.10726492 6.817353 1.0000000 Confirmed
#> 28  C00079 2.472516  2.504132  0.29238281 4.006911 0.8897796 Confirmed
#> 29  C00148 3.525716  3.544549  2.15580502 4.881016 1.0000000 Confirmed
#> 30  C00149 2.900391  2.927963  0.83580244 4.412906 0.9619238 Confirmed
#> 31  C03406 2.345437  2.402859  0.13279458 4.128212 0.8376754 Confirmed
#> 32  C03413 3.620795  3.626360  1.99222144 5.081519 0.9959920 Confirmed
#> 33  C03793 2.496885  2.524761  0.44275584 3.846496 0.8917836 Confirmed
#> 34  C00153 2.439765  2.458623 -0.63621922 3.816896 0.8997996 Confirmed
#> 35  C13482 1.877529  1.933530 -0.39910700 3.356259 0.6232465 Confirmed
#> 36  C00022 4.284624  4.292924  2.36990325 5.884062 1.0000000 Confirmed
#> 37  C00255 3.714586  3.724309  2.11511926 5.438547 1.0000000 Confirmed
#> 38  C00447 2.410840  2.441581  0.15100445 3.791734 0.8657315 Confirmed
#> 39  C00670 1.996932  2.016001 -0.13423363 3.297750 0.6913828 Confirmed
#> 40  C00042 3.501690  3.514203  1.93453430 4.856591 0.9959920 Confirmed
#> 41  C00167 2.429689  2.479067 -1.00100150 3.838921 0.8877756 Confirmed
#> 42  C00106 1.800091  1.880362 -1.00100150 3.237648 0.5851703 Confirmed