Feature selection using lasso or elastic network

Feature selection using lasso or elastic network

Usage

ML_alpha(object, method = "lasso")

Arguments

object

A dataframe-like data object containing log-metabolite intensity values, with columns corresponding to metabolites and must containing the group column, and the rows corresponding to the samples

method

the feature selection method,default is "lasso", "elastic" is optional

Value

test

Examples

library(dplyr)

meta_dat1 <- t(meta_dat) %>%
  as.data.frame() %>%
  dplyr::mutate(group=group)
result_ML_lasso <- ML_alpha(meta_dat1, method="lasso")
#> Loading required package: ggplot2
#> Loading required package: lattice
result_ML_lasso
#>   feature          s1
#> 1  C02045 0.321816672
#> 2  C00022 0.002814505