change the compound name to refmet name

if the compound name can not be changed to kegg id, then the compound name is changed to refmet id

Usage

name2refmet(metabolites_name)

Arguments

metabolites_name

the compound name need to changed to refmet name

Value

test

Examples

compound_name <- c("2-Hydroxybutyric acid","1-Methyladenosine","tt","2-Aminooctanoic acid")
name2refmet(compound_name)
#> Warning: cannot remove file 'test.data.1', reason 'No such file or directory'
#> Warning: cannot remove file 'test.data.2', reason 'No such file or directory'
#> Warning: cannot remove file 'test.data.3', reason 'No such file or directory'
#> Warning: cannot remove file 'test.data.4', reason 'No such file or directory'
#>              Input_name           Refmet_name    Formula   Super_class
#> 1 2-Hydroxybutyric acid 2-Hydroxybutyric acid     C4H8O3   Fatty Acyls
#> 2     1-Methyladenosine     1-Methyladenosine C11H15N5O4 Nucleic acids
#> 3                    tt                    tt          -             -
#> 4  2-Aminooctanoic acid  2-Aminocaprylic acid   C8H17NO2   Fatty Acyls
#>    Main_class              Sub_class
#> 1 Fatty acids             Hydroxy FA
#> 2     Purines Purine ribonucleosides
#> 3           -                      -
#> 4 Fatty acids               Amino FA