Function to view enrichment results

xEnrichViewer is supposed to view results of enrichment analysis.

Usage

xEnrichViewer(
  eTerm,
  top_num = 10,
  sortBy = c("adjp", "fdr", "pvalue", "zscore", "fc", "nAnno", "nOverlap", "or", "none"),
  decreasing = NULL,
  details = FALSE
)

Arguments

eTerm

an object of class "eTerm"

top_num

the number of the top terms (sorted according to 'sortBy' below) will be viewed

sortBy

which statistics will be used for sorting and viewing gene sets (terms). It can be "adjp" or "fdr" for adjusted p value (FDR), "pvalue" for p value, "zscore" for enrichment z-score, "fc" for enrichment fold change, "nAnno" for the number of sets (terms), "nOverlap" for the number in overlaps, "or" for the odds ratio, and "none" for ordering according to ID of terms

decreasing

logical to indicate whether to sort in a decreasing order. If it is null, it would be true for "zscore", "nAnno" or "nOverlap"; otherwise it would be false

details

logical to indicate whether the detailed information of gene sets (terms) is also viewed. By default, it sets to false for no inclusion

Value

a data frame with following components:

id: term ID; as rownames

name: term name

nAnno: number in members annotated by a term

nOverlap: number in overlaps

fc: enrichment fold changes

zscore: enrichment z-score

pvalue: nominal p value

adjp: adjusted p value (FDR)

or: a vector containing odds ratio

CIl: a vector containing lower bound confidence interval for the odds ratio

CIu: a vector containing upper bound confidence interval for the odds ratio

distance: term distance or other information; optional, it is only appended when "details" is true

members_Overlap: members (represented as Gene Symbols) in overlaps; optional, it is only appended when "details" is true

members_Anno: members (represented as Gene Symbols) in annotations; optional, it is only appended when "details" is true

Note

none

See also

Examples

RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
if (FALSE) {
xEnrichViewer(eTerm)
}