Feature selection using Boruta

Feature selection using Boruta

Usage

ML_Boruta(object)

Arguments

object

A dataframe-like data object containing log-metabolite intensity values, with columns corresponding to metabolites and must containing the group column, and the rows corresponding to the samples

Value

test

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

meta_dat1 <- t(meta_dat) %>%
  as.data.frame() %>%
  dplyr::mutate(group=group)

result_ML_Boruta <- ML_Boruta(meta_dat1)
#>  1. run of importance source...
#>  2. run of importance source...
#>  3. run of importance source...
#>  4. run of importance source...
#>  5. run of importance source...
#>  6. run of importance source...
#>  7. run of importance source...
#>  8. run of importance source...
#>  9. run of importance source...
#>  10. run of importance source...
#>  11. run of importance source...
#>  12. run of importance source...
#>  13. run of importance source...
#>  14. run of importance source...
#>  15. run of importance source...
#> After 15 iterations, +0.54 secs: 
#>  confirmed 20 attributes: C00022, C00025, C00042, C00073, C00148 and 15 more;
#>  rejected 121 attributes: C00002, C00003, C00006, C00008, C00015 and 116 more;
#>  still have 78 attributes left.
#>  16. run of importance source...
#>  17. run of importance source...
#>  18. run of importance source...
#>  19. run of importance source...
#> After 19 iterations, +0.69 secs: 
#>  rejected 13 attributes: C00016, C00020, C00021, C00035, C00055 and 8 more;
#>  still have 65 attributes left.
#>  20. run of importance source...
#>  21. run of importance source...
#>  22. run of importance source...
#>  23. run of importance source...
#> After 23 iterations, +0.8 secs: 
#>  confirmed 1 attribute: C00490;
#>  rejected 5 attributes: C00043, C00078, C00093, C02862, C05401;
#>  still have 59 attributes left.
#>  24. run of importance source...
#>  25. run of importance source...
#>  26. run of importance source...
#> After 26 iterations, +0.87 secs: 
#>  confirmed 4 attributes: C00079, C00153, C00570, C05673;
#>  rejected 3 attributes: C00152, C00188, C03740;
#>  still have 52 attributes left.
#>  27. run of importance source...
#>  28. run of importance source...
#>  29. run of importance source...
#>  30. run of importance source...
#> After 30 iterations, +0.96 secs: 
#>  confirmed 1 attribute: C00167;
#>  rejected 2 attributes: C00004, C00791;
#>  still have 49 attributes left.
#>  31. run of importance source...
#>  32. run of importance source...
#>  33. run of importance source...
#> After 33 iterations, +1 secs: 
#>  confirmed 3 attributes: C00149, C00170, C00447;
#>  rejected 2 attributes: C00183, C00800;
#>  still have 44 attributes left.
#>  34. run of importance source...
#>  35. run of importance source...
#>  36. run of importance source...
#> After 36 iterations, +1.1 secs: 
#>  confirmed 2 attributes: C00307, C03406;
#>  rejected 2 attributes: C00064, C11924;
#>  still have 40 attributes left.
#>  37. run of importance source...
#>  38. run of importance source...
#>  39. run of importance source...
#>  40. run of importance source...
#>  41. run of importance source...
#>  42. run of importance source...
#>  43. run of importance source...
#>  44. run of importance source...
#>  45. run of importance source...
#>  46. run of importance source...
#>  47. run of importance source...
#>  48. run of importance source...
#> After 48 iterations, +1.5 secs: 
#>  rejected 2 attributes: C00158, C00624;
#>  still have 38 attributes left.
#>  49. run of importance source...
#>  50. run of importance source...
#> After 50 iterations, +1.5 secs: 
#>  rejected 2 attributes: C00385, C00438;
#>  still have 36 attributes left.
#>  51. run of importance source...
#>  52. run of importance source...
#>  53. run of importance source...
#>  54. run of importance source...
#>  55. run of importance source...
#>  56. run of importance source...
#> After 56 iterations, +1.6 secs: 
#>  rejected 1 attribute: C05100;
#>  still have 35 attributes left.
#>  57. run of importance source...
#>  58. run of importance source...
#>  59. run of importance source...
#> After 59 iterations, +1.7 secs: 
#>  confirmed 2 attributes: C00013, C00097;
#>  still have 33 attributes left.
#>  60. run of importance source...
#>  61. run of importance source...
#> After 61 iterations, +1.8 secs: 
#>  confirmed 1 attribute: C00352;
#>  rejected 1 attribute: C01419;
#>  still have 31 attributes left.
#>  62. run of importance source...
#>  63. run of importance source...
#>  64. run of importance source...
#>  65. run of importance source...
#>  66. run of importance source...
#>  67. run of importance source...
#>  68. run of importance source...
#>  69. run of importance source...
#> After 69 iterations, +1.9 secs: 
#>  confirmed 2 attributes: C00010, C06192;
#>  still have 29 attributes left.
#>  70. run of importance source...
#>  71. run of importance source...
#>  72. run of importance source...
#> After 72 iterations, +2 secs: 
#>  rejected 1 attribute: C00037;
#>  still have 28 attributes left.
#>  73. run of importance source...
#>  74. run of importance source...
#>  75. run of importance source...
#> After 75 iterations, +2.1 secs: 
#>  rejected 1 attribute: C00408;
#>  still have 27 attributes left.
#>  76. run of importance source...
#>  77. run of importance source...
#>  78. run of importance source...
#>  79. run of importance source...
#>  80. run of importance source...
#> After 80 iterations, +2.2 secs: 
#>  rejected 1 attribute: C16513;
#>  still have 26 attributes left.
#>  81. run of importance source...
#>  82. run of importance source...
#>  83. run of importance source...
#>  84. run of importance source...
#>  85. run of importance source...
#>  86. run of importance source...
#>  87. run of importance source...
#>  88. run of importance source...
#>  89. run of importance source...
#>  90. run of importance source...
#>  91. run of importance source...
#>  92. run of importance source...
#>  93. run of importance source...
#>  94. run of importance source...
#>  95. run of importance source...
#>  96. run of importance source...
#>  97. run of importance source...
#>  98. run of importance source...
#>  99. run of importance source...
#>  100. run of importance source...
#>  101. run of importance source...
#>  102. run of importance source...
#>  103. run of importance source...
#>  104. run of importance source...
#>  105. run of importance source...
#>  106. run of importance source...
#>  107. run of importance source...
#>  108. run of importance source...
#>  109. run of importance source...
#>  110. run of importance source...
#>  111. run of importance source...
#>  112. run of importance source...
#>  113. run of importance source...
#>  114. run of importance source...
#>  115. run of importance source...
#>  116. run of importance source...
#>  117. run of importance source...
#> After 117 iterations, +3 secs: 
#>  confirmed 1 attribute: C00106;
#>  still have 25 attributes left.
#>  118. run of importance source...
#>  119. run of importance source...
#>  120. run of importance source...
#>  121. run of importance source...
#>  122. run of importance source...
#>  123. run of importance source...
#>  124. run of importance source...
#>  125. run of importance source...
#>  126. run of importance source...
#>  127. run of importance source...
#>  128. run of importance source...
#>  129. run of importance source...
#>  130. run of importance source...
#>  131. run of importance source...
#>  132. run of importance source...
#>  133. run of importance source...
#>  134. run of importance source...
#>  135. run of importance source...
#>  136. run of importance source...
#>  137. run of importance source...
#>  138. run of importance source...
#>  139. run of importance source...
#>  140. run of importance source...
#>  141. run of importance source...
#>  142. run of importance source...
#>  143. run of importance source...
#>  144. run of importance source...
#>  145. run of importance source...
#>  146. run of importance source...
#>  147. run of importance source...
#>  148. run of importance source...
#>  149. run of importance source...
#>  150. run of importance source...
#>  151. run of importance source...
#>  152. run of importance source...
#>  153. run of importance source...
#>  154. run of importance source...
#>  155. run of importance source...
#>  156. run of importance source...
#>  157. run of importance source...
#>  158. run of importance source...
#>  159. run of importance source...
#>  160. run of importance source...
#>  161. run of importance source...
#> After 161 iterations, +3.9 secs: 
#>  confirmed 1 attribute: C00670;
#>  still have 24 attributes left.
#>  162. run of importance source...
#>  163. run of importance source...
#>  164. run of importance source...
#>  165. run of importance source...
#>  166. run of importance source...
#>  167. run of importance source...
#>  168. run of importance source...
#>  169. run of importance source...
#>  170. run of importance source...
#>  171. run of importance source...
#>  172. run of importance source...
#>  173. run of importance source...
#>  174. run of importance source...
#>  175. run of importance source...
#>  176. run of importance source...
#>  177. run of importance source...
#>  178. run of importance source...
#>  179. run of importance source...
#>  180. run of importance source...
#>  181. run of importance source...
#>  182. run of importance source...
#>  183. run of importance source...
#>  184. run of importance source...
#>  185. run of importance source...
#>  186. run of importance source...
#>  187. run of importance source...
#>  188. run of importance source...
#>  189. run of importance source...
#>  190. run of importance source...
#>  191. run of importance source...
#>  192. run of importance source...
#>  193. run of importance source...
#>  194. run of importance source...
#>  195. run of importance source...
#>  196. run of importance source...
#>  197. run of importance source...
#>  198. run of importance source...
#>  199. run of importance source...
#>  200. run of importance source...
#>  201. run of importance source...
#>  202. run of importance source...
#>  203. run of importance source...
#>  204. run of importance source...
#>  205. run of importance source...
#>  206. run of importance source...
#>  207. run of importance source...
#>  208. run of importance source...
#>  209. run of importance source...
#>  210. run of importance source...
#>  211. run of importance source...
#>  212. run of importance source...
#>  213. run of importance source...
#>  214. run of importance source...
#>  215. run of importance source...
#>  216. run of importance source...
#>  217. run of importance source...
#>  218. run of importance source...
#>  219. run of importance source...
#>  220. run of importance source...
#>  221. run of importance source...
#>  222. run of importance source...
#>  223. run of importance source...
#>  224. run of importance source...
#>  225. run of importance source...
#>  226. run of importance source...
#>  227. run of importance source...
#>  228. run of importance source...
#>  229. run of importance source...
#>  230. run of importance source...
#>  231. run of importance source...
#>  232. run of importance source...
#>  233. run of importance source...
#>  234. run of importance source...
#>  235. run of importance source...
#>  236. run of importance source...
#>  237. run of importance source...
#>  238. run of importance source...
#>  239. run of importance source...
#>  240. run of importance source...
#>  241. run of importance source...
#>  242. run of importance source...
#>  243. run of importance source...
#>  244. run of importance source...
#>  245. run of importance source...
#>  246. run of importance source...
#>  247. run of importance source...
#>  248. run of importance source...
#>  249. run of importance source...
#>  250. run of importance source...
#>  251. run of importance source...
#> After 251 iterations, +5.5 secs: 
#>  rejected 1 attribute: C00065;
#>  still have 23 attributes left.
#>  252. run of importance source...
#>  253. run of importance source...
#>  254. run of importance source...
#>  255. run of importance source...
#>  256. run of importance source...
#>  257. run of importance source...
#>  258. run of importance source...
#>  259. run of importance source...
#>  260. run of importance source...
#>  261. run of importance source...
#>  262. run of importance source...
#>  263. run of importance source...
#>  264. run of importance source...
#>  265. run of importance source...
#>  266. run of importance source...
#>  267. run of importance source...
#>  268. run of importance source...
#>  269. run of importance source...
#>  270. run of importance source...
#>  271. run of importance source...
#>  272. run of importance source...
#>  273. run of importance source...
#>  274. run of importance source...
#>  275. run of importance source...
#>  276. run of importance source...
#>  277. run of importance source...
#>  278. run of importance source...
#>  279. run of importance source...
#>  280. run of importance source...
#>  281. run of importance source...
#>  282. run of importance source...
#>  283. run of importance source...
#>  284. run of importance source...
#>  285. run of importance source...
#>  286. run of importance source...
#>  287. run of importance source...
#>  288. run of importance source...
#>  289. run of importance source...
#>  290. run of importance source...
#>  291. run of importance source...
#>  292. run of importance source...
#>  293. run of importance source...
#>  294. run of importance source...
#>  295. run of importance source...
#>  296. run of importance source...
#>  297. run of importance source...
#>  298. run of importance source...
#>  299. run of importance source...
#>  300. run of importance source...
#>  301. run of importance source...
#>  302. run of importance source...
#>  303. run of importance source...
#>  304. run of importance source...
#>  305. run of importance source...
#>  306. run of importance source...
#>  307. run of importance source...
#>  308. run of importance source...
#>  309. run of importance source...
#>  310. run of importance source...
#>  311. run of importance source...
#>  312. run of importance source...
#>  313. run of importance source...
#>  314. run of importance source...
#>  315. run of importance source...
#>  316. run of importance source...
#>  317. run of importance source...
#>  318. run of importance source...
#>  319. run of importance source...
#>  320. run of importance source...
#>  321. run of importance source...
#>  322. run of importance source...
#>  323. run of importance source...
#>  324. run of importance source...
#>  325. run of importance source...
#>  326. run of importance source...
#>  327. run of importance source...
#>  328. run of importance source...
#>  329. run of importance source...
#>  330. run of importance source...
#>  331. run of importance source...
#>  332. run of importance source...
#>  333. run of importance source...
#>  334. run of importance source...
#>  335. run of importance source...
#>  336. run of importance source...
#>  337. run of importance source...
#>  338. run of importance source...
#>  339. run of importance source...
#>  340. run of importance source...
#>  341. run of importance source...
#>  342. run of importance source...
#>  343. run of importance source...
#>  344. run of importance source...
#>  345. run of importance source...
#>  346. run of importance source...
#>  347. run of importance source...
#>  348. run of importance source...
#>  349. run of importance source...
#>  350. run of importance source...
#>  351. run of importance source...
#>  352. run of importance source...
#>  353. run of importance source...
#>  354. run of importance source...
#>  355. run of importance source...
#>  356. run of importance source...
#>  357. run of importance source...
#>  358. run of importance source...
#>  359. run of importance source...
#>  360. run of importance source...
#>  361. run of importance source...
#>  362. run of importance source...
#>  363. run of importance source...
#>  364. run of importance source...
#>  365. run of importance source...
#>  366. run of importance source...
#>  367. run of importance source...
#>  368. run of importance source...
#>  369. run of importance source...
#>  370. run of importance source...
#>  371. run of importance source...
#>  372. run of importance source...
#>  373. run of importance source...
#>  374. run of importance source...
#>  375. run of importance source...
#>  376. run of importance source...
#>  377. run of importance source...
#>  378. run of importance source...
#>  379. run of importance source...
#>  380. run of importance source...
#>  381. run of importance source...
#>  382. run of importance source...
#>  383. run of importance source...
#>  384. run of importance source...
#> After 384 iterations, +7.9 secs: 
#>  rejected 1 attribute: C00186;
#>  still have 22 attributes left.
#>  385. run of importance source...
#>  386. run of importance source...
#>  387. run of importance source...
#>  388. run of importance source...
#> After 388 iterations, +8 secs: 
#>  confirmed 1 attribute: C13482;
#>  still have 21 attributes left.
#>  389. run of importance source...
#>  390. run of importance source...
#>  391. run of importance source...
#>  392. run of importance source...
#>  393. run of importance source...
#>  394. run of importance source...
#>  395. run of importance source...
#>  396. run of importance source...
#>  397. run of importance source...
#>  398. run of importance source...
#>  399. run of importance source...
#>  400. run of importance source...
#>  401. run of importance source...
#>  402. run of importance source...
#>  403. run of importance source...
#>  404. run of importance source...
#>  405. run of importance source...
#>  406. run of importance source...
#>  407. run of importance source...
#>  408. run of importance source...
#>  409. run of importance source...
#>  410. run of importance source...
#>  411. run of importance source...
#>  412. run of importance source...
#>  413. run of importance source...
#>  414. run of importance source...
#>  415. run of importance source...
#>  416. run of importance source...
#>  417. run of importance source...
#>  418. run of importance source...
#>  419. run of importance source...
#>  420. run of importance source...
#>  421. run of importance source...
#>  422. run of importance source...
#>  423. run of importance source...
#>  424. run of importance source...
#>  425. run of importance source...
#>  426. run of importance source...
#>  427. run of importance source...
#>  428. run of importance source...
#>  429. run of importance source...
#>  430. run of importance source...
#>  431. run of importance source...
#>  432. run of importance source...
#>  433. run of importance source...
#>  434. run of importance source...
#>  435. run of importance source...
#>  436. run of importance source...
#>  437. run of importance source...
#>  438. run of importance source...
#>  439. run of importance source...
#>  440. run of importance source...
#>  441. run of importance source...
#> After 441 iterations, +8.9 secs: 
#>  confirmed 1 attribute: C03264;
#>  still have 20 attributes left.
#>  442. run of importance source...
#>  443. run of importance source...
#>  444. run of importance source...
#>  445. run of importance source...
#>  446. run of importance source...
#>  447. run of importance source...
#>  448. run of importance source...
#>  449. run of importance source...
#>  450. run of importance source...
#>  451. run of importance source...
#>  452. run of importance source...
#>  453. run of importance source...
#>  454. run of importance source...
#>  455. run of importance source...
#>  456. run of importance source...
#>  457. run of importance source...
#>  458. run of importance source...
#>  459. run of importance source...
#> After 459 iterations, +9.2 secs: 
#>  confirmed 1 attribute: C00601;
#>  still have 19 attributes left.
#>  460. run of importance source...
#>  461. run of importance source...
#>  462. run of importance source...
#>  463. run of importance source...
#>  464. run of importance source...
#>  465. run of importance source...
#>  466. run of importance source...
#>  467. run of importance source...
#>  468. run of importance source...
#>  469. run of importance source...
#>  470. run of importance source...
#>  471. run of importance source...
#>  472. run of importance source...
#>  473. run of importance source...
#>  474. run of importance source...
#>  475. run of importance source...
#>  476. run of importance source...
#>  477. run of importance source...
#>  478. run of importance source...
#>  479. run of importance source...
#>  480. run of importance source...
#>  481. run of importance source...
#>  482. run of importance source...
#>  483. run of importance source...
#>  484. run of importance source...
#>  485. run of importance source...
#>  486. run of importance source...
#>  487. run of importance source...
#>  488. run of importance source...
#>  489. run of importance source...
#>  490. run of importance source...
#>  491. run of importance source...
#>  492. run of importance source...
#>  493. run of importance source...
#>  494. run of importance source...
#>  495. run of importance source...
#>  496. run of importance source...
#>  497. run of importance source...
#>  498. run of importance source...
#>  499. run of importance source...
result_ML_Boruta
#>       name  meanImp medianImp      minImp   maxImp  normHits  decision
#> 1   C03264 1.807909  1.890535 -1.00100150 3.192239 0.6052104 Confirmed
#> 2   C02630 3.904201  3.907530  2.23148770 5.273588 1.0000000 Confirmed
#> 3   C00170 2.320137  2.383305  0.03629764 3.677534 0.8496994 Confirmed
#> 4   C18170 3.504635  3.485503  1.99657024 4.950293 0.9979960 Confirmed
#> 5   C06192 2.226020  2.252420 -0.15535207 3.605125 0.8136273 Confirmed
#> 6   C00267 4.249434  4.265181  2.71962194 5.732876 1.0000000 Confirmed
#> 7   C05378 3.372497  3.400252  1.00100150 4.580840 0.9979960 Confirmed
#> 8   C00307 2.431422  2.465864  0.38834251 3.890010 0.8677355 Confirmed
#> 9   C00570 2.377377  2.425383  0.67975787 4.025020 0.8737475 Confirmed
#> 10  C05673 2.059550  2.120350 -1.00100150 3.575686 0.7114228 Confirmed
#> 11  C05674 4.205909  4.245613  2.58794798 5.596388 1.0000000 Confirmed
#> 12  C00010 2.174915  2.176662 -0.54707777 3.729528 0.7735471 Confirmed
#> 13  C00352 1.960835  1.993343  0.40702693 3.375289 0.6873747 Confirmed
#> 14  C00013 2.418982  2.458271 -0.06025382 3.989990 0.8597194 Confirmed
#> 15  C00365 3.986083  3.989186  2.65324954 5.022842 1.0000000 Confirmed
#> 16  C00242 3.865565  3.867330  1.64777354 5.141176 1.0000000 Confirmed
#> 17  C00490 2.567933  2.609922  0.88561200 3.729620 0.9238477 Confirmed
#> 18 C05938  3.784330  3.779723  1.97520352 5.319923 0.9959920 Confirmed
#> 19  C02291 4.342211  4.343853  2.54753180 6.195052 1.0000000 Confirmed
#> 20  C00097 2.104833  2.160514 -0.44730307 3.626908 0.7595190 Confirmed
#> 21  C02045 5.110925  5.131432  2.60233105 6.487123 1.0000000 Confirmed
#> 22  C02985 3.736901  3.755093  2.15106877 5.022963 1.0000000 Confirmed
#> 23  C00025 3.821855  3.828091  2.04405647 5.132712 1.0000000 Confirmed
#> 24  C00407 3.495047  3.506900  2.00369990 4.711387 0.9959920 Confirmed
#> 25  C00073 5.108951  5.155963  2.74064717 6.408177 1.0000000 Confirmed
#> 26  C00079 2.437029  2.460363  0.84391892 3.742544 0.8937876 Confirmed
#> 27  C00148 3.453748  3.478979  2.00903725 4.781362 0.9959920 Confirmed
#> 28  C00149 2.836733  2.851271  0.62012980 4.104433 0.9519038 Confirmed
#> 29  C03406 2.351558  2.409619 -0.42844806 3.709005 0.8517034 Confirmed
#> 30  C03413 3.629044  3.636868  2.11701431 5.051383 1.0000000 Confirmed
#> 31  C03793 2.488445  2.520756 -0.20814671 3.811534 0.8957916 Confirmed
#> 32  C00153 2.415698  2.463387  0.06726758 3.640149 0.8717435 Confirmed
#> 33  C00601 1.793065  1.851176 -0.39858311 3.323290 0.5911824 Confirmed
#> 34  C13482 1.814811  1.883224 -1.61006779 3.097376 0.6112224 Confirmed
#> 35  C00022 4.245764  4.271655  2.81372115 5.689834 1.0000000 Confirmed
#> 36  C00255 3.766215  3.775998  2.42772492 5.136435 1.0000000 Confirmed
#> 37  C00447 2.364507  2.376125 -0.79140096 3.842949 0.8717435 Confirmed
#> 38  C00670 1.913253  1.923820 -1.00100150 3.101893 0.6613226 Confirmed
#> 39  C00042 3.502316  3.514991  1.95045560 4.721383 1.0000000 Confirmed
#> 40  C00167 2.357317  2.370685 -0.19598929 3.843674 0.8577154 Confirmed
#> 41  C00106 1.847609  1.908517 -0.84142528 3.113480 0.6352705 Confirmed