feature selection using Boruta

feature selection using Boruta

Usage

ML_Boruta(object)

Arguments

object

A dataframe-like data object containing log-metabolite intensity values, with columns corresponding to metabolites and must containing the group column, and the rows corresponding to the samples

Value

test

Examples

library(dplyr)
meta_dat1 <- t(meta_dat) %>%
  as.data.frame() %>%
  dplyr::mutate(group=group)
result_ML_Boruta <- ML_Boruta(meta_dat1)
#>  1. run of importance source...
#>  2. run of importance source...
#>  3. run of importance source...
#>  4. run of importance source...
#>  5. run of importance source...
#>  6. run of importance source...
#>  7. run of importance source...
#>  8. run of importance source...
#>  9. run of importance source...
#>  10. run of importance source...
#>  11. run of importance source...
#>  12. run of importance source...
#>  13. run of importance source...
#>  14. run of importance source...
#>  15. run of importance source...
#> After 15 iterations, +0.46 secs: 
#>  confirmed 19 attributes: C00022, C00025, C00042, C00073, C00148 and 14 more;
#>  rejected 121 attributes: C00002, C00003, C00006, C00008, C00015 and 116 more;
#>  still have 79 attributes left.
#>  16. run of importance source...
#>  17. run of importance source...
#>  18. run of importance source...
#>  19. run of importance source...
#> After 19 iterations, +0.53 secs: 
#>  confirmed 2 attributes: C00490, C03793;
#>  rejected 10 attributes: C00043, C00064, C00117, C00178, C00214 and 5 more;
#>  still have 67 attributes left.
#>  20. run of importance source...
#>  21. run of importance source...
#>  22. run of importance source...
#>  23. run of importance source...
#> After 23 iterations, +0.59 secs: 
#>  rejected 8 attributes: C00004, C00112, C00158, C00183, C00429 and 3 more;
#>  still have 59 attributes left.
#>  24. run of importance source...
#>  25. run of importance source...
#>  26. run of importance source...
#> After 26 iterations, +0.64 secs: 
#>  confirmed 4 attributes: C00149, C00153, C00307, C00570;
#>  rejected 9 attributes: C00035, C00037, C00152, C00188, C00295 and 4 more;
#>  still have 46 attributes left.
#>  27. run of importance source...
#>  28. run of importance source...
#>  29. run of importance source...
#>  30. run of importance source...
#>  31. run of importance source...
#>  32. run of importance source...
#>  33. run of importance source...
#> After 33 iterations, +0.75 secs: 
#>  confirmed 4 attributes: C00079, C00167, C00170, C00447;
#>  still have 42 attributes left.
#>  34. run of importance source...
#>  35. run of importance source...
#>  36. run of importance source...
#> After 36 iterations, +0.79 secs: 
#>  rejected 2 attributes: C00105, C00137;
#>  still have 40 attributes left.
#>  37. run of importance source...
#>  38. run of importance source...
#>  39. run of importance source...
#> After 39 iterations, +0.84 secs: 
#>  rejected 1 attribute: C00588;
#>  still have 39 attributes left.
#>  40. run of importance source...
#>  41. run of importance source...
#>  42. run of importance source...
#> After 42 iterations, +0.88 secs: 
#>  confirmed 2 attributes: C00013, C05673;
#>  rejected 1 attribute: C00624;
#>  still have 36 attributes left.
#>  43. run of importance source...
#>  44. run of importance source...
#>  45. run of importance source...
#> After 45 iterations, +0.93 secs: 
#>  rejected 1 attribute: C02642;
#>  still have 35 attributes left.
#>  46. run of importance source...
#>  47. run of importance source...
#>  48. run of importance source...
#> After 48 iterations, +0.96 secs: 
#>  rejected 1 attribute: C00800;
#>  still have 34 attributes left.
#>  49. run of importance source...
#>  50. run of importance source...
#>  51. run of importance source...
#>  52. run of importance source...
#>  53. run of importance source...
#>  54. run of importance source...
#>  55. run of importance source...
#>  56. run of importance source...
#>  57. run of importance source...
#>  58. run of importance source...
#>  59. run of importance source...
#> After 59 iterations, +1.1 secs: 
#>  confirmed 1 attribute: C06192;
#>  rejected 1 attribute: C00385;
#>  still have 32 attributes left.
#>  60. run of importance source...
#>  61. run of importance source...
#> After 61 iterations, +1.1 secs: 
#>  confirmed 1 attribute: C03406;
#>  still have 31 attributes left.
#>  62. run of importance source...
#>  63. run of importance source...
#>  64. run of importance source...
#>  65. run of importance source...
#>  66. run of importance source...
#>  67. run of importance source...
#>  68. run of importance source...
#>  69. run of importance source...
#>  70. run of importance source...
#>  71. run of importance source...
#>  72. run of importance source...
#> After 72 iterations, +1.3 secs: 
#>  confirmed 1 attribute: C00010;
#>  still have 30 attributes left.
#>  73. run of importance source...
#>  74. run of importance source...
#>  75. run of importance source...
#>  76. run of importance source...
#>  77. run of importance source...
#>  78. run of importance source...
#>  79. run of importance source...
#>  80. run of importance source...
#>  81. run of importance source...
#>  82. run of importance source...
#>  83. run of importance source...
#>  84. run of importance source...
#>  85. run of importance source...
#>  86. run of importance source...
#>  87. run of importance source...
#>  88. run of importance source...
#>  89. run of importance source...
#>  90. run of importance source...
#>  91. run of importance source...
#>  92. run of importance source...
#>  93. run of importance source...
#>  94. run of importance source...
#>  95. run of importance source...
#> After 95 iterations, +1.5 secs: 
#>  confirmed 1 attribute: C00097;
#>  still have 29 attributes left.
#>  96. run of importance source...
#>  97. run of importance source...
#>  98. run of importance source...
#>  99. run of importance source...
#>  100. run of importance source...
#>  101. run of importance source...
#>  102. run of importance source...
#>  103. run of importance source...
#> After 103 iterations, +1.6 secs: 
#>  rejected 1 attribute: C01419;
#>  still have 28 attributes left.
#>  104. run of importance source...
#>  105. run of importance source...
#>  106. run of importance source...
#>  107. run of importance source...
#>  108. run of importance source...
#> After 108 iterations, +1.7 secs: 
#>  confirmed 1 attribute: C00670;
#>  still have 27 attributes left.
#>  109. run of importance source...
#>  110. run of importance source...
#>  111. run of importance source...
#>  112. run of importance source...
#>  113. run of importance source...
#>  114. run of importance source...
#>  115. run of importance source...
#>  116. run of importance source...
#>  117. run of importance source...
#>  118. run of importance source...
#>  119. run of importance source...
#>  120. run of importance source...
#> After 120 iterations, +1.9 secs: 
#>  rejected 1 attribute: C00065;
#>  still have 26 attributes left.
#>  121. run of importance source...
#>  122. run of importance source...
#>  123. run of importance source...
#>  124. run of importance source...
#>  125. run of importance source...
#>  126. run of importance source...
#>  127. run of importance source...
#>  128. run of importance source...
#>  129. run of importance source...
#>  130. run of importance source...
#> After 130 iterations, +2 secs: 
#>  confirmed 1 attribute: C00352;
#>  still have 25 attributes left.
#>  131. run of importance source...
#>  132. run of importance source...
#> After 132 iterations, +2 secs: 
#>  rejected 1 attribute: C00186;
#>  still have 24 attributes left.
#>  133. run of importance source...
#>  134. run of importance source...
#>  135. run of importance source...
#>  136. run of importance source...
#>  137. run of importance source...
#>  138. run of importance source...
#>  139. run of importance source...
#>  140. run of importance source...
#>  141. run of importance source...
#>  142. run of importance source...
#>  143. run of importance source...
#>  144. run of importance source...
#>  145. run of importance source...
#>  146. run of importance source...
#>  147. run of importance source...
#>  148. run of importance source...
#>  149. run of importance source...
#>  150. run of importance source...
#>  151. run of importance source...
#>  152. run of importance source...
#>  153. run of importance source...
#>  154. run of importance source...
#>  155. run of importance source...
#>  156. run of importance source...
#>  157. run of importance source...
#>  158. run of importance source...
#>  159. run of importance source...
#>  160. run of importance source...
#>  161. run of importance source...
#>  162. run of importance source...
#>  163. run of importance source...
#>  164. run of importance source...
#>  165. run of importance source...
#>  166. run of importance source...
#>  167. run of importance source...
#>  168. run of importance source...
#>  169. run of importance source...
#>  170. run of importance source...
#>  171. run of importance source...
#>  172. run of importance source...
#>  173. run of importance source...
#>  174. run of importance source...
#>  175. run of importance source...
#>  176. run of importance source...
#>  177. run of importance source...
#>  178. run of importance source...
#>  179. run of importance source...
#>  180. run of importance source...
#>  181. run of importance source...
#>  182. run of importance source...
#>  183. run of importance source...
#>  184. run of importance source...
#>  185. run of importance source...
#>  186. run of importance source...
#>  187. run of importance source...
#>  188. run of importance source...
#>  189. run of importance source...
#>  190. run of importance source...
#>  191. run of importance source...
#>  192. run of importance source...
#>  193. run of importance source...
#>  194. run of importance source...
#>  195. run of importance source...
#>  196. run of importance source...
#>  197. run of importance source...
#>  198. run of importance source...
#> After 198 iterations, +2.8 secs: 
#>  confirmed 1 attribute: C05851;
#>  still have 23 attributes left.
#>  199. run of importance source...
#>  200. run of importance source...
#>  201. run of importance source...
#>  202. run of importance source...
#>  203. run of importance source...
#>  204. run of importance source...
#>  205. run of importance source...
#>  206. run of importance source...
#>  207. run of importance source...
#>  208. run of importance source...
#>  209. run of importance source...
#>  210. run of importance source...
#>  211. run of importance source...
#>  212. run of importance source...
#>  213. run of importance source...
#>  214. run of importance source...
#>  215. run of importance source...
#>  216. run of importance source...
#>  217. run of importance source...
#>  218. run of importance source...
#>  219. run of importance source...
#>  220. run of importance source...
#>  221. run of importance source...
#> After 221 iterations, +3.1 secs: 
#>  confirmed 1 attribute: C13482;
#>  still have 22 attributes left.
#>  222. run of importance source...
#>  223. run of importance source...
#>  224. run of importance source...
#>  225. run of importance source...
#>  226. run of importance source...
#>  227. run of importance source...
#>  228. run of importance source...
#>  229. run of importance source...
#>  230. run of importance source...
#>  231. run of importance source...
#>  232. run of importance source...
#>  233. run of importance source...
#>  234. run of importance source...
#>  235. run of importance source...
#>  236. run of importance source...
#>  237. run of importance source...
#>  238. run of importance source...
#>  239. run of importance source...
#>  240. run of importance source...
#>  241. run of importance source...
#>  242. run of importance source...
#> After 242 iterations, +3.4 secs: 
#>  rejected 1 attribute: C16513;
#>  still have 21 attributes left.
#>  243. run of importance source...
#>  244. run of importance source...
#>  245. run of importance source...
#>  246. run of importance source...
#>  247. run of importance source...
#>  248. run of importance source...
#>  249. run of importance source...
#>  250. run of importance source...
#>  251. run of importance source...
#>  252. run of importance source...
#>  253. run of importance source...
#>  254. run of importance source...
#>  255. run of importance source...
#>  256. run of importance source...
#>  257. run of importance source...
#>  258. run of importance source...
#>  259. run of importance source...
#>  260. run of importance source...
#>  261. run of importance source...
#>  262. run of importance source...
#>  263. run of importance source...
#>  264. run of importance source...
#>  265. run of importance source...
#>  266. run of importance source...
#>  267. run of importance source...
#>  268. run of importance source...
#>  269. run of importance source...
#>  270. run of importance source...
#>  271. run of importance source...
#>  272. run of importance source...
#>  273. run of importance source...
#>  274. run of importance source...
#>  275. run of importance source...
#>  276. run of importance source...
#>  277. run of importance source...
#>  278. run of importance source...
#>  279. run of importance source...
#>  280. run of importance source...
#>  281. run of importance source...
#>  282. run of importance source...
#>  283. run of importance source...
#>  284. run of importance source...
#>  285. run of importance source...
#>  286. run of importance source...
#>  287. run of importance source...
#>  288. run of importance source...
#>  289. run of importance source...
#>  290. run of importance source...
#>  291. run of importance source...
#>  292. run of importance source...
#>  293. run of importance source...
#>  294. run of importance source...
#>  295. run of importance source...
#>  296. run of importance source...
#>  297. run of importance source...
#>  298. run of importance source...
#>  299. run of importance source...
#>  300. run of importance source...
#>  301. run of importance source...
#>  302. run of importance source...
#>  303. run of importance source...
#>  304. run of importance source...
#>  305. run of importance source...
#>  306. run of importance source...
#>  307. run of importance source...
#>  308. run of importance source...
#>  309. run of importance source...
#>  310. run of importance source...
#>  311. run of importance source...
#>  312. run of importance source...
#> After 312 iterations, +4.2 secs: 
#>  confirmed 1 attribute: C00106;
#>  still have 20 attributes left.
#>  313. run of importance source...
#>  314. run of importance source...
#>  315. run of importance source...
#>  316. run of importance source...
#>  317. run of importance source...
#>  318. run of importance source...
#>  319. run of importance source...
#>  320. run of importance source...
#>  321. run of importance source...
#>  322. run of importance source...
#>  323. run of importance source...
#>  324. run of importance source...
#>  325. run of importance source...
#>  326. run of importance source...
#>  327. run of importance source...
#>  328. run of importance source...
#> After 328 iterations, +4.4 secs: 
#>  confirmed 1 attribute: C00719;
#>  still have 19 attributes left.
#>  329. run of importance source...
#>  330. run of importance source...
#>  331. run of importance source...
#>  332. run of importance source...
#>  333. run of importance source...
#>  334. run of importance source...
#>  335. run of importance source...
#>  336. run of importance source...
#>  337. run of importance source...
#>  338. run of importance source...
#>  339. run of importance source...
#>  340. run of importance source...
#>  341. run of importance source...
#>  342. run of importance source...
#>  343. run of importance source...
#>  344. run of importance source...
#>  345. run of importance source...
#>  346. run of importance source...
#>  347. run of importance source...
#>  348. run of importance source...
#>  349. run of importance source...
#>  350. run of importance source...
#>  351. run of importance source...
#>  352. run of importance source...
#>  353. run of importance source...
#>  354. run of importance source...
#>  355. run of importance source...
#>  356. run of importance source...
#>  357. run of importance source...
#>  358. run of importance source...
#>  359. run of importance source...
#>  360. run of importance source...
#>  361. run of importance source...
#>  362. run of importance source...
#>  363. run of importance source...
#>  364. run of importance source...
#>  365. run of importance source...
#>  366. run of importance source...
#>  367. run of importance source...
#>  368. run of importance source...
#> After 368 iterations, +4.9 secs: 
#>  rejected 1 attribute: C06241;
#>  still have 18 attributes left.
#>  369. run of importance source...
#>  370. run of importance source...
#>  371. run of importance source...
#>  372. run of importance source...
#>  373. run of importance source...
#>  374. run of importance source...
#>  375. run of importance source...
#>  376. run of importance source...
#>  377. run of importance source...
#>  378. run of importance source...
#>  379. run of importance source...
#>  380. run of importance source...
#>  381. run of importance source...
#>  382. run of importance source...
#>  383. run of importance source...
#>  384. run of importance source...
#>  385. run of importance source...
#>  386. run of importance source...
#>  387. run of importance source...
#>  388. run of importance source...
#>  389. run of importance source...
#>  390. run of importance source...
#>  391. run of importance source...
#>  392. run of importance source...
#>  393. run of importance source...
#>  394. run of importance source...
#>  395. run of importance source...
#>  396. run of importance source...
#>  397. run of importance source...
#>  398. run of importance source...
#>  399. run of importance source...
#>  400. run of importance source...
#>  401. run of importance source...
#>  402. run of importance source...
#> After 402 iterations, +5.3 secs: 
#>  rejected 1 attribute: C00051;
#>  still have 17 attributes left.
#>  403. run of importance source...
#>  404. run of importance source...
#>  405. run of importance source...
#>  406. run of importance source...
#>  407. run of importance source...
#>  408. run of importance source...
#>  409. run of importance source...
#>  410. run of importance source...
#>  411. run of importance source...
#>  412. run of importance source...
#>  413. run of importance source...
#>  414. run of importance source...
#>  415. run of importance source...
#>  416. run of importance source...
#>  417. run of importance source...
#>  418. run of importance source...
#>  419. run of importance source...
#>  420. run of importance source...
#>  421. run of importance source...
#>  422. run of importance source...
#>  423. run of importance source...
#>  424. run of importance source...
#>  425. run of importance source...
#>  426. run of importance source...
#>  427. run of importance source...
#>  428. run of importance source...
#>  429. run of importance source...
#>  430. run of importance source...
#>  431. run of importance source...
#>  432. run of importance source...
#>  433. run of importance source...
#>  434. run of importance source...
#>  435. run of importance source...
#> After 435 iterations, +5.7 secs: 
#>  confirmed 1 attribute: C03264;
#>  still have 16 attributes left.
#>  436. run of importance source...
#>  437. run of importance source...
#>  438. run of importance source...
#>  439. run of importance source...
#>  440. run of importance source...
#>  441. run of importance source...
#>  442. run of importance source...
#>  443. run of importance source...
#>  444. run of importance source...
#>  445. run of importance source...
#>  446. run of importance source...
#>  447. run of importance source...
#>  448. run of importance source...
#>  449. run of importance source...
#>  450. run of importance source...
#>  451. run of importance source...
#>  452. run of importance source...
#>  453. run of importance source...
#>  454. run of importance source...
#>  455. run of importance source...
#>  456. run of importance source...
#>  457. run of importance source...
#>  458. run of importance source...
#>  459. run of importance source...
#>  460. run of importance source...
#>  461. run of importance source...
#>  462. run of importance source...
#>  463. run of importance source...
#>  464. run of importance source...
#>  465. run of importance source...
#>  466. run of importance source...
#>  467. run of importance source...
#>  468. run of importance source...
#>  469. run of importance source...
#>  470. run of importance source...
#>  471. run of importance source...
#>  472. run of importance source...
#>  473. run of importance source...
#>  474. run of importance source...
#>  475. run of importance source...
#>  476. run of importance source...
#>  477. run of importance source...
#>  478. run of importance source...
#>  479. run of importance source...
#>  480. run of importance source...
#>  481. run of importance source...
#>  482. run of importance source...
#>  483. run of importance source...
#>  484. run of importance source...
#>  485. run of importance source...
#>  486. run of importance source...
#>  487. run of importance source...
#>  488. run of importance source...
#>  489. run of importance source...
#>  490. run of importance source...
#>  491. run of importance source...
#>  492. run of importance source...
#>  493. run of importance source...
#>  494. run of importance source...
#>  495. run of importance source...
#>  496. run of importance source...
#>  497. run of importance source...
#>  498. run of importance source...
#>  499. run of importance source...